refactor: --tune-only を main() から別モジュールへ出す
main() が 1,049 行あった。AGENTS.md の規約は関数 50 行なので 20 倍。

--tune-only のブロックが最大の自己完結した塊だった。args.tune_only の
ときだけ走り、必ず return するので、後続の処理はこのブロックが計算した
ものに何も依存していない。240 行をそのまま tuning.rs へ移した。

中身は 1 行も書き換えていない（インデントのみ）。呼び出しは

  if args.tune_only {
      return tuning::run_probe(&args, &opt);
  }

外部依存は args と opt の 2 つだけで、どちらも変更しないので参照で渡す。
最初 cfg と threads も渡していたが、正規表現ベースの解析が過剰に拾って
いただけで実際には使っていなかったため落とした。

main.rs 1,819 -> 1,584 行、main() 1,049 -> 809 行。

## ここで止める理由

残り 412 行は走査と出力で、外部変数を 10 個参照する。関数に切り出すと
引数 10 個になるので、文脈を持つ構造体の新設が要る。

一方でテストの状況が以下だった。

  hyperdu-cli のテスト        5 件
  --top / --json / --csv の
  出力経路のテスト            0 件

無テストのコードを、公開直前に構造体を新設しながら組み替えるのは順序が
逆になる。先に出力の特性テストを足してから分割する。

## 検証

- cargo test --workspace: 220 件通過 / 16 スイート
- cargo clippy --workspace --all-targets: 警告 0
- --tune-only を実行し、移動前と同じ推奨コマンドを出すことを確認
